+++ src/lib/util/minionIsBusy.ts
@@ -1,0 +1,6 @@
+import { getActivityOfUser } from '../settings/settings';
+
+export function minionIsBusy(userID: bigint | string): boolean {
+	const usersTask = getActivityOfUser(userID.toString());
+	return Boolean(usersTask);
+}
