+++ src/mahoji/lib/abstracted_commands/farmingContractCommand.ts
@@ -3,11 +3,11 @@ import { CommandResponse } from 'mahoji/dist/lib/structures/ICommand';
 
 import { client } from '../../..';
 import { Favours, gotFavour } from '../../../lib/minions/data/kourendFavour';
-import { defaultFarmingContract } from '../../../lib/minions/farming';
+import { defaultFarmingContract, getPlantToGrow } from '../../../lib/minions/farming';
 import { ContractOption, FarmingContract, FarmingContractDifficultyLevel } from '../../../lib/minions/farming/types';
-import { getPlantToGrow } from '../../../lib/skilling/functions/calcFarmingContracts';
 import { newChatHeadImage } from '../../../lib/util/chatHeadImage';
-import { minionIsBusy, minionName } from '../../../lib/util/minionUtils';
+import { minionIsBusy } from '../../../lib/util/minionIsBusy';
+import { minionName } from '../../../lib/util/minionUtils';
 import { findPlant } from '../../commands/farming';
 import { getMahojiBank, getSkillsOfMahojiUser, mahojiUserSettingsUpdate } from '../../mahojiSettings';
